Move __from_arrow_ to base class
